% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailies.R
\docType{package}
\name{dailies}
\alias{dailies}
\title{dailies: A package for working with the RStudio daily build pages.}
\description{
RStudio offer their daily builds of the IDE to the community so that we can
help out with bug reports and check out new features before they hit the
stable releases. The RStudio dailies website is available at
\url{https://dailies.rstudio.com}
}
\details{
The dailies package provides several functions for working with the dailies
page itself and is intended to make it easier for people to get useful URLs
from within R.
}
\section{Important Note}{
 Some functions request data from the RStudio
  website. Please be respectful in your usage of these functions. The dailies
  site is offered as a service to the community and we should not repay that
  kindness by hammering the site.\cr This package is very much "unofficial".
}

\section{dailies functions}{
 The most useful function is
  \code{latest()} which returns the url for the latest version.
}

