% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Get the dailies page URL}
\usage{
get_url(os, type = "server")
}
\arguments{
\item{os}{The operating system}

\item{type}{One of "server" (default) or "desktop"}
}
\value{
The URL of the dailies page
}
\description{
Returns the URL of the dailies page for a given os and type combination.
}
\details{
The following os and type combinations are supported:\cr
type = "server"
\itemize{
\item ubuntu (currently v16 and up)
\item debian
\item redhat
\item suse
}
type = "desktop"
\itemize{
\item windows
\item macos
\item ubuntu
\item fedora
}
}
