% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{latest daily download URL}
\usage{
latest(os, type = "desktop")
}
\arguments{
\item{os}{The operating system}

\item{type}{One of "desktop" (default) or "server"}
}
\value{
The URL of the latest daily build
}
\description{
Returns the most recent daily version download URL for a given os and type
The options available are derived directly from the API.
}
