% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_downloads.R
\name{list_downloads}
\alias{list_downloads}
\title{List available downloads}
\usage{
list_downloads(os, type = "server")
}
\arguments{
\item{os}{The operating system}

\item{type}{One of "server" (default) or "desktop"}
}
\value{
A character vector of download URLs
}
\description{
Returns a list of the available downloads for a given os and type combination
}
