% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{latest daily download URL}
\usage{
latest(os, type = "server")
}
\arguments{
\item{os}{The operating system}

\item{type}{One of "server" (default) or "desktop"}
}
\value{
The most recent download URL
}
\description{
Returns the most recent daily version download URL for a given os and type
}
